move test case base classes to separate module
